src/analysis_engine/analyzer.py
import logging
from typing import Dict, Any, List
from datetime import datetime
import numpy as np
from sklearn.ensemble import IsolationForest

class TraceAnalyzer:
    """Analyzes system traces for potential security threats."""
    
    def __init__(self):
        self.logger = logging.getLogger(__name__)
        self.anomaly_detector = IsolationForest(
            contamination=0.1,
            random_state=42
        )
        self.baseline_data = []
    
    def analyze_traces(self, traces: Dict[str, Any]) -> Dict[str, Any]:
        """Analyze collected system traces for anomalies."""
        analysis_result = {
            'timestamp': datetime.now().isoformat(),
            'anomalies': self._detect_anomalies(traces),
            'risk_assessment': self._assess_risk(traces),
            'recommendations': self._generate_recommendations(traces)
        }
        return analysis_result
    
    def _detect_anomalies(self, traces: Dict[str, Any]) -> List[Dict[str, Any]]:
        """Detect anomalies in system traces using isolation forest."""
        features = self._extract_features(traces)
        if not self.baseline_data:
            self.baseline_data = features
            return []
        
        # Fit and predict anomalies
        self.anomaly_detector.fit(self.baseline_data)
        anomaly_scores = self.anomaly_detector.score_samples([features])
        
        return [{
            'score': float(anomaly_scores[0]),
            'threshold': -0.5,
            'is_anomaly': anomaly_scores[0] < -0.5
        }]
    
    def _extract_features(self, traces: Dict[str, Any]) -> np.ndarray:
        """Extract numerical features from traces for anomaly detection."""
        features = [
            traces['system_resources']['cpu_percent'],
            traces['system_resources']['memory']['percent'],
            len(traces['network_connections']),
            len(traces['processes'])
        ]
        return np.array(features).reshape(1, -1)
    
    def _assess_risk(self, traces: Dict[str, Any]) -> Dict[str, Any]:
        """Assess the risk level based on system traces."""
        risk_factors = {
            'unusual_processes': self._check_unusual_processes(traces),
            'network_anomalies': self._check_network_anomalies(traces),
            'resource_usage': self._check_resource_usage(traces)
        }
        
        risk_level = self._calculate_risk_level(risk_factors)
        return {
            'level': risk_level,
            'factors': risk_factors
        }
    
    def _check_unusual_processes(self, traces: Dict[str, Any]) -> List[Dict[str, Any]]:
        """Check for unusual processes."""
        unusual = []
        for proc in traces['processes']:
            if self._is_suspicious_process(proc):
                unusual.append(proc)
        return unusual
    
    def _check_network_anomalies(self, traces: Dict[str, Any]) -> List[Dict[str, Any]]:
        """Check for network connection anomalies."""
        anomalies = []
        for conn in traces['network_connections']:
            if self._is_suspicious_connection(conn):
                anomalies.append(conn)
        return anomalies
    
    def _check_resource_usage(self, traces: Dict[str, Any]) -> Dict[str, Any]:
        """Check for unusual resource usage patterns."""
        return {
            'high_cpu': traces['system_resources']['cpu_percent'] > 90,
            'high_memory': traces['system_resources']['memory']['percent'] > 90
        }
    
    def _calculate_risk_level(self, risk_factors: Dict[str, Any]) -> str:
        """Calculate overall risk level."""
        if len(risk_factors['unusual_processes']) > 5 or len(risk_factors['network_anomalies']) > 3:
            return 'HIGH'
        elif len(risk_factors['unusual_processes']) > 2 or len(risk_factors['network_anomalies']) > 1:
            return 'MEDIUM'
        return 'LOW'
    
    def _is_suspicious_process(self, process: Dict[str, Any]) -> bool:
        """Check if a process appears suspicious."""
        # Implement more sophisticated detection logic
        return False
    
    def _is_suspicious_connection(self, connection: Dict[str, Any]) -> bool:
        """Check if a network connection appears suspicious."""
        # Implement more sophisticated detection logic
        return False
    
    def _generate_recommendations(self, traces: Dict[str, Any]) -> List[str]:
        """Generate security recommendations based on analysis."""
        recommendations = []
        
        # Add basic recommendations
        if traces['system_resources']['cpu_percent'] > 90:
            recommendations.append("Investigate processes causing high CPU usage")
        if traces['system_resources']['memory']['percent'] > 90:
            recommendations.append("Monitor memory usage and potential memory leaks")
        
        return recommendations
