requirements.txt
fastapi
uvicorn
numpy
scikit-learn
matplotlib
psutil
transformers
torch
lief
networkx
pandas
requests
pyyaml
datasets
